requirements.txt
fastapi
uvicorn
httpx
pydantic-ai
mem0ai
requests
python-dotenv>=1.0.0 